+++ lib/core/routing/app_router.dart
@@ -3,6 +3,7 @@ import 'package:thimar/core/routing/routes.dart';
 import 'package:thimar/features/auth/login/presentation/login_screen.dart';
 import 'package:thimar/features/splash/presentation/splash_screen.dart';
 
+import '../../features/auth/forget_password/presentation/forget_password_screen.dart';
 import '../../features/auth/register/presentation/register_screen.dart';
 
 class AppRouter {
@@ -18,6 +19,9 @@ class AppRouter {
         );case Routes.registerScreen:
         return MaterialPageRoute(
           builder: (_) => RegisterScreen(),
+        );case Routes.forgetPasswordScreen:
+        return MaterialPageRoute(
+          builder: (_) => ForgetPasswordScreen(),
         );
     }
     return null;
