+++ lib/core/routing/routes.dart
@@ -2,4 +2,5 @@ abstract class Routes{
   static const String splashScreen = '/SplashScreen';
   static const String loginScreen = '/LoginScreen';
   static const String registerScreen = '/RegisterScreen';
+  static const String forgetPasswordScreen = '/ForgetPasswordScreen';
 }
