+++ lib/features/auth/login/presentation/widgets/login_text_forms_widget.dart
@@ -7,6 +7,7 @@ import '../../../../../core/theming/app_colors.dart';
 import '../../../../../core/theming/app_styles.dart';
 import '../../../../../core/utils/spacing.dart';
 import '../../../../../core/widgets/app_text_form_field.dart';
+import '../../../widgets/auth_phone_and_country_widget.dart';
 
 class LoginTextFormsWidget extends StatelessWidget {
   const LoginTextFormsWidget({super.key});
@@ -15,66 +16,7 @@ class LoginTextFormsWidget extends StatelessWidget {
   Widget build(BuildContext context) {
     return Column(
       children: [
-        Row(
-          children: [
-            Flexible(
-              flex: 1,
-              child: Container(
-                height: 50.h,
-                padding: EdgeInsets.symmetric(horizontal: 12.w),
-                decoration: BoxDecoration(
-                  border: Border.all(
-                    color: AppColors.lightGrayColor,
-                    width: 1.w,
-                  ),
-                  borderRadius: BorderRadius.circular(16.r),
-                ),
-                child: Center(
-                  child: CountryCodePicker(
-                    onChanged: _onCountryChange,
-                    initialSelection: 'SA',
-                    favorite: ['+966', 'SA'],
-                    showCountryOnly: false,
-                    showOnlyCountryWhenClosed: false,
-                    alignLeft: false,
-                    builder: (countryCode) {
-                      return Column(
-                        mainAxisAlignment: MainAxisAlignment.center,
-                        crossAxisAlignment: CrossAxisAlignment.center,
-                        children: [
-                          Image.asset(
-                            countryCode!.flagUri!,
-                            package: 'country_code_picker',
-                            width: 24.w,
-                            height: 18.h,
-                          ),
-                          horizontalSpace(6),
-                          Text(
-                            countryCode.dialCode ?? "",
-                            style: AppStyles.font16MediumBold,
-                          ),
-                        ],
-                      );
-                    },
-                  ),
-                ),
-              ),
-            ),
-            horizontalSpace(8),
-            Flexible(
-              flex: 4,
-              child: AppTextFormField(
-                hintText: "رقم الجوال",
-                validator: (value) {},
-                prefixIcon: Image.asset(
-                  AppAssets.phoneImage,
-                  width: 22.w,
-                  height: 22.h,
-                ),
-              ),
-            ),
-          ],
-        ),
+        AuthPhoneAndCountryWidget(),
         verticalSpace(16),
         AppTextFormField(
           hintText: "كلمة المرور",
@@ -90,7 +32,5 @@ class LoginTextFormsWidget extends StatelessWidget {
     );
   }
 
-  void _onCountryChange(CountryCode countryCode) {
-    print("New Country selected: " + countryCode.toString());
-  }
+
 }
